+++ src/constants.rs
@@ -1,0 +1,9 @@
+pub const BOARD_WIDTH: f32 = 264.0;
+pub const BOARD_HEIGHT: f32 = 462.0;
+pub const SCREEN_WIDTH: f32 = 792.0;
+pub const SCREEN_HEIGHT: f32 = BOARD_HEIGHT + BOARD_HEIGHT / 4.0;
+pub const FALL_TIME: f32 = 1.0;
+pub const POINTS_FOR_ONE_LINE: u32 = 40;
+pub const POINTS_FOR_TWO_LINES: u32 = 100;
+pub const POINTS_FOR_THREE_LINES: u32 = 300;
+pub const POINTS_FOR_MORE_THAN_THREE_LINES: u32 = 1200;
